tests: Add tests of specs and README back in
This patch adds the README and specs tests back in. The tests were temporarily
removed when the Quilt specs were moved to the GitHub Quilt organization.

Since we do not want `make check` to depend on internet access, and since there
is no easy way to have govendor exclude certain tests, the spec tests do not use
Go's `testing` API. The specification tests will instead be run separately with
the command `make check-specs`.