ps: Use image name in spec for deployed Dockerfiles
Before, `quilt ps` when run on a deployment containing Dockerfiles built
in the cluster would print the munged image name stored in the database
on Workers. This image name is an implementation detail, and not
relevant to users, so we now simply print the image name stored in the
leader's database.