quiltctl: Add debug-logs command
When developing quilt, we often have a need to see logs from several
different machines and containers, but it's a pain to log into each
and extract the logs we need. This patch adds a command that fetches
logs from a set of machines and containers and stores them in files
on the user's machine.

This has the nice side effect of giving users an easy way to get
logs that they can send us in the case that they're encountering
bugs that they can't figure out.