docker: Create parent directories when writing files
Before, specifying a `filepathToContent` only worked when the parent
directory of the `filepath` already existed. This was because we simply
uploaded a tarball containing the filename to the parent directory. For
example, the path `/a/b/c/d` would be uploaded as a tarball containing
`d` at path `/a/b/c/`.

We now encode the entire directory structure within the tarball, and upload to `/`
if the path is an absolute path, and `.` otherwise. For example, `/a/b/c`
would be uploaded as a tarball containing `a/b/c` at path `/`, and
`a/b` would be uploaded as a tarball containing `a/b` at path `.`.