registry: Don't block transaction while building images
Before, we built Docker images inside the transaction that queries them
to ensure that the image specification doesn't change. However, because
building containers takes awhile, this blocked other parts of the system
that depended on the image table (e.g. the scheduler) for
inappropriately long periods of time.

We now build the images outside the transaction, and then assign the
image IDs to image rows that are still applicable.

This should fix the CI log test. It was failing because the master was
losing its leadership while the registry blocked the image table, which
blocked the scheduler, which acquired a lock for the etcd table in the
same transaction as the image table.  When the master lost its
leadership, it removed `ovn-northd`, causing errors throughout the
cluster.