scheduler: Don't block etcd table while scheduling containers
Before, we unnecessarily held the lock for the etcd table for the entire
duration of scheduling. This is excessive because we simply need the
etcd table to check whether we are the leader.

Note that the previous code that held the lock for the entire
transaction seemed to imply that we could not lose our leadership
because we were blocking the etcd table. This is incorrect because if
the elector does not run while we are blocking the etcd table, it will
still be unable to refresh the leader node, and thus lose its
leadership.