quilt-tester: Don't hardcode wait time for spec completion
Before, we waited 5 minutes per test for the containers in the test spec
to boot. This made test runs unnecessarily long because most specs did
not require 5 minutes to boot, and caused the MEAN test to fail because
building the MEAN Dockerfile took too long.

We now wait up to 10 minutes, but exit early if all the containers in
the spec have booted before then.