minion: Preserve SSH keys installed by cloud config
Before, we immediately started syncing the SSH keys as soon as the
minion booted. This resulted in a brief period where no keys were
installed. This is because Minion.AuthorizedKeys was not populated until
the foreman connected; therefore, before the foreman sends the
MinionConfig over, we would overwrite the keys installed by the cloud
config. To alleviate this, we do not touch the keys until the foreman
has sent us some keys to install.