debug-logs: Default to putting the logs in a tarball
A common use case for the `debug-logs` command is to have users send
us their logs so we can debug issues. In order to make that simpler,
the `debug-logs` command defaults to storing all the logs in a
tarball instead of a folder. If desired, the logs can be left stored
in a folder by passing the `-tar=false` flag.