amazon: Don't include pending instances in `List`
Before, we would include pending reserved machines in `List`. This
resulted in a bug where the cluster would prematurely try to update
floating IPs on instances that had not yet fully booted.

We now block machine boot and stops until they are no longer pending.
If we time out before machines transition to the "running" state, we
destroy them to avoid leaking machines that might not ever properly
transition.

Co-Authored-By: Ethan J. Jackson <ejj@eecs.berkeley.edu>