quilt-tester: Increase wait time after detecting containers starting
With the move to Jenkins, a timing issue caused network tests to fail.
It was discovered that these errors were resolved when the test was
ran again at a later time, so the tester has been modified to wait
a little bit longer before running tests on a set of containers.