debug-logs: Simplify expected files handling in unit tests
The previous implementation of the debug-logs unit tests only checked
that expected files were present, and that unexpected files were not
present. This was cumbersome because we had to specify two fields.
Furthermore, it was easy to write tests that ignored extra files
that weren't explicitly specified in `expNotFiles`, or missing files
that weren't specified in `expFiles`.