google: Explicitly define credentials source
Before, we obtained an OAuth client for the Google API by setting an
environment variable, then obtaining a default client that referenced
this environment variable when attempting to obtain a client.

We now simply read in the JSON config file, and use that to create a
client. This is easier to follow, and will furthermore allow us to parse
information from the JSON config in a future patch.