nat: Ignore order of options when determining equality
Iptables doesn't preserve rule options in the same order that they are
inserted. For example, we may add the following rule:

-s 10.89.157.197/32 -p tcp -m tcp --dport 80 -o eth0 -j MASQUERADE

But when we listing, the rule would display as:

-s 10.89.157.197/32 -o eth0 -p tcp -m tcp --dport 80 -j MASQUERADE

This resulted in our POSTROUTING rules being re-added every
iteration because the different options ordering caused our code to
believe the current rules were incorrect, and needed to be replaced.

We now sort the options deterministically when comparing rules.