cloudcfg: Don't start minion before loading OVS modules
Before, we inserted the OVS kernel modules and ran the minion in
parallel. This resulted in a possible race condition where we try
to create quilt-int before we have the necessary modules.

This race condition would be particularly likely if the kernel module
had to be built at runtime before insertion.