quilt: Don't evaluate Javascript in non-stitch unit tests
Before, some unit tests that weren't testing the language itself used
the Javascript, rather than the deployment object, to setup the test.
Moving away from the Javascript for unit tests that don't require it
makes the language much easier to modify, as there are less tests to fix
when breaking changes are made.