network: Create exclusive port mappings in the engine
Before, we created exclusive port mappings while compiling specs.
However, not being able to have two containers listen on the same port
on the same machine is an attribute of the deployment engine
implementation, not of the language.

This simplification of the language implementation will also slightly
simplify the transition to Node.js.