stitch: Implement language as a Node.js module
Before, we used Otto (a Go implementation of the Javascript
specification) to evaluate specs entirely within Go. This patch
migrates the Javascript evaluation to occur within the Node runtime.

This has several benefits:
- Node has been around awhile, and is thus much more mature than Otto.
It's maintained better, and it has developed stories and tooling that we
won't have to reinvent (such as `npm` for package management, and `use
strict`).
- Node is familiar to many developers, so the learning curve for writing
Quilt specs will be learning how to use the Quilt module, rather than
learning the Quilt language.
- Node has an existing ecosystem of modules that Quilt users can use.

Co-Authored-By: Kevin Lin <kaikai526@gmail.com>