travis: Autodeploy master release using tags
Before, we tried to autodeploy the release for master using a
condition that matches all tags, and the master branch. However, when
Travis deploys branches, it generates a tag name based on the commit ID.
This is inconvenient because the name is not human-friendly, and changes
with each new commit to master.

We now simply update the `dev` tag for each commit in master, and Travis
builds all tags.