docs: Fix export commands
The export commands listed in the installation part of the README used single
quotes, which caused environment variables to not be evaluated. We now use
double quotes to properly expand variables. For consistency, and because double
quotes are safer than no quotes, the same changes were made to
GettingStarted.md.