version: Change default version name to `dev`
This change makes the image associated with the master branch have the
tag `dev`. This is better because it matches the tag we use to track the
HEAD of the master branch. Furthermore, this will cleanly fix a build bug
introduced by `travis: Push Docker images for tags rather than branches`
because we won't require special logic for the `master` branch.