tests: Set working directory for readme/spec tests
Setting the working directory is usually done automatically done
by the go testing framework, but we need to do it ourselves for
these tests, which we run using "go run". This ensures that, if
one test changes the working directory, it won't affect the other
tests.

This commit sets the working directory to be the package directory,
for consistency with go's testing framework.