network: Explicitly check whether the logical switch exists
Before, we didn't check the result of `CreateLogicalSwitch` because it
returned an error if the logical switch already existed.

This commit adds a `LogicalSwitchExists` method so that we can call
`CreateLogicalSwitch` only when necessary. This is helpful because an
error is then never the expected behavior, so we can properly log it.