stitch: Rename connect to allowFrom
The name "connect" has been a source of confusion because it's
not obvious which direction the connection is being allowed in, so
this commit changes connect to allowFrom (where users used
a.connect(5, b), they should now use b.allowFrom(a, 5)). This is
intended to be more intuitive to use. This commit warns users of
connect that they should use the newer allowFrom (but doesn't yet
delete it, which we can do in a future release).

This commit also removes Connection from the public exports from
bindings.js, because this shouldn't be used by users directly.

This commit does not update the docs, which will be done as part
of the larger docs re-write.