command: Add hook for command setup and teardown
This commit adds the `BeforeRun` and `AfterRun` functions to the `SubCommand`
interface to facilitate common transformations of command line flags.
For example, it is extremely common for subcommands to obtain an API
client connected to the daemon. Rather than replicate this logic in the
`Main` function of each subcommand, this patch will allow subcommands to
utilize one `connection` object that obtains the client during the
`BeforeRun` phase, and closes it during `AfterRun`.