stitch: Consistently use single quotes
We should use consistent style in our JavaScript. Eventually
we should switch to a style guide and linter, but in the meantime,
this change makes an incremental improvement towards consistency by
changing all strings to use single quotes or back quotes, which is
what's suggested by the Google JavaScript style guide.