digitalocean: Temporarily decrease number of supported regions
Because of the current cluster implementation, we list machines in
every region. To decrease the number of requests, this commit decreases
the number of regions that DigitalOcean tracks. This commit will be made
obsolete once we modify the cluster to only track regions that are
explicitly specified by the user.