client: Simplify error return logic
This small change makes it explicit that the returned error at this
point cannot be non-nil, because there is already a guard clause a
couple lines up. Thus, any code that mutates the error can safely do so
within the guard clause, and doesn't have to worry about errors thrown
after that block.