cluster: Rename instance to launchLoc
cluster.go uses "instance" to refer to one region for a provider.
This name is confusing because "instance" is used elsewhere to
mean a virtual machine that a provider has allocated to Quilt,
so this commit renames it to "launchLoc".

This commit also makes a few minor logging changes to log both
the provider name and the region, instead of just the name.