quiltctl: Proxy API client connections through daemon
This patch modifies the daemon such that it proxies the
`QueryContainers`, `QueryLabels`, and `QueryConnections` RPCs to the
cluster. This way, client-users don't have to worry about getting worker
clients or the leader client. This will be especially useful once we've
enforced TLS, as the keys can be used by the daemon, and not managed by
all clients.