minion: Add flag to specify NAT rule interfaces
We currently restrict the NAT rules for inbound and outbound connections
to only apply to the interface associated with the default route.
However, the default route heuristic doesn't work for Vagrant --
Vagrant has separate interfaces for inbound and outbound connections.

This commit adds support for a command line option to specify what
interface the NAT rules should apply to. If an interface isn't
explicitly provided, the default route heuristic is used.