cluster: Simplify the interface for getting a machine's role
Before, the cluster obtained the role for the minion running on a
machine by passing a slice of machines, which the foreman modified
with the minion roles, and returned. However, the only field that the
foreman accessed was the machine's public IP, which we now pass in
directly as an argument.

This has the additional benefit that in an upcoming patch, the foreman
won't require access to the private `joinMachine` struct defined in the
cluster.

This commit also fixes the unit tests to properly follow the expected
codepath where the foreman reports the machine's role to the cluster.
The previous tests simply used the `Role` specified at boot.