cluster: Remove joining fields from `machine.Machine`
Before, `machine.Machine` contained fields (specifically Provider,
Region, and Role) that didn't have much meaning to the providers, but
were useful for joining. Furthermore, the cluster required that
providers correctly set the Provider and Region during `List`s.

These attributes are known by the cluster; therefore, the providers
should not have to set them. To make the `machine.Machine` struct
easier to follow, and to simplify the provider code, this commit
introduces a `joinMachine` struct that wraps `machine.Machine` in
addition to fields for `provider`, `region`, and `role`.

Note that `Role` was not completely removed from the `machine.Machine`
struct because it's needed to generate the cloud config. That will be
done in a future patch.