vagrant: Refuse to boot preemptible machines
Preemptible machines are not defined on Vagrant, so this commit modifies
the Vagrant provider to throw an error when requested to boot them.
Otherwise, the cluster would continuously create and destroy machines
because Vagrant's `List()` would return machines with the `preemptible`
flag set to false, thus causing the cluster join to not match them.